% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{evalAugW}
\alias{evalAugW}
\title{.evalAugW
calls TMLE to use super learner to evalute preliminary predictions for
Q(0,W) and Q(1,W) conditioning on stage 1 covariates}
\usage{
evalAugW(Y, A, W, Delta, id, family, SL.library)
}
\arguments{
\item{Y}{outcome vector}

\item{A}{binary treatment indicator}

\item{W}{covariate matrix}

\item{Delta}{outcome missingness indicator}

\item{id}{identifier of i.i.d. unit}

\item{family}{outcome regression family}

\item{SL.library}{super learner library for outcome regression modeling}
}
\value{
\code{W.Q}, nx2 matrix of outcome predictions based on stage 1
covariates
}
\description{
.evalAugW
calls TMLE to use super learner to evalute preliminary predictions for
Q(0,W) and Q(1,W) conditioning on stage 1 covariates
}
