% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printUtils.R
\name{print.summary.twoStageTMLE}
\alias{print.summary.twoStageTMLE}
\title{print.summary.twoStageTMLE}
\usage{
\method{print}{summary.twoStageTMLE}(x, ...)
}
\arguments{
\item{x}{an object of class summary.twoStageTMLE}

\item{...}{additional arguments (i)}
}
\value{
print object
}
\description{
print.summary.twoStageTMLE
}
