% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printUtils.R
\name{print.twoStageTMLE}
\alias{print.twoStageTMLE}
\title{print.twoStageTMLE}
\usage{
\method{print}{twoStageTMLE}(x, ...)
}
\arguments{
\item{x}{an object of class twoStageTMLE}

\item{...}{additional arguments (i)}
}
\value{
print tmle results using print.tmle
method from tmle package
}
\description{
print.twoStageTMLE
}
