% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{setV}
\alias{setV}
\title{Utilities
setV
Set the number of cross-validation
folds as a function of effective sample size
See Phillips 2023 doi.org/10.1093/ije/dyad023}
\usage{
setV(n.effective)
}
\arguments{
\item{n.effective}{the effective sample size}
}
\value{
the number of cross-validation folds
}
\description{
Utilities
setV
Set the number of cross-validation
folds as a function of effective sample size
See Phillips 2023 doi.org/10.1093/ije/dyad023
}
