% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printUtils.R
\name{summary.twoStage}
\alias{summary.twoStage}
\title{summary.twoStageTMLE}
\usage{
\method{summary}{twoStage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{twoStageTMLE}}

\item{...}{Other arguments passed to the tmle function in the tmle package}
}
\value{
A list containing the missingness model, terms, coefficients, type,
}
\description{
Summarizes estimation procedure for missing 2nd stage covariates
}
