% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printUtils.R
\name{summary.twoStageTMLE}
\alias{summary.twoStageTMLE}
\title{summary.twoStageTMLE}
\usage{
\method{summary}{twoStageTMLE}(object, ...)
}
\arguments{
\item{object}{an object of class twoStageTMLE}

\item{...}{additional arguments (ignored)}
}
\value{
list summarizing the two-stage procedure components,
summary of the twoStage missingness estimation
summary of the tmle for estimating the parameter
}
\description{
summary.twoStageTMLE
}
