% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{twoStageDesignTMLENews}
\alias{twoStageDesignTMLENews}
\title{twoStageDesignTMLENews
Get news about recent updates and bug fixes}
\usage{
twoStageDesignTMLENews(...)
}
\arguments{
\item{...}{ignored}
}
\value{
invisible character string giving the path to the file found.
}
\description{
twoStageDesignTMLENews
Get news about recent updates and bug fixes
}
